requirements.txt
pillow==11.2.1
pandas==2.2.3
Flask==3.1.0

firebase-admin==6.7.0
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-api-python-client==2.166.0
google-auth==2.38.0
cloudinary==1.43.0
python-dotenv==1.1.0
requests
numpy==1.26.4
openpyxl==3.1.5
pyyaml==6.0.2
onnxruntime==1.22.1
gunicorn
gdown
opencv-python